// my-app/src/app/services/evento/evento-interface.ts
export interface EventoInterface {
    id?: number;
    title?: string;
    description?: string;
    start?: any;
    start_time?: any;
    end?: any;
    end_time?: any;
    duration?: number;
    preference_id?: number;
    preference_name?: string;
    office_id?: number;
    office_name?: string;
    user_id?: number;
    user_name?: string;
    created_at?: string;
    updated_at?: string;
}